s3/replication/requirements.txt
#:runtime-requirements:

flake8==3.9.1
aiohttp==3.7.4.post0
pyYaml==5.4.1
